pyproject.toml
[project]
name = "ml-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "harish gehlot", email = "gehloth03@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "hydra-core>=1.3.2",
    "joblib>=1.5.2",
    "pandera>=0.27.0",
    "polars>=1.36.1",
    "pre-commit>=4.5.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.9",
    "rust-just>=1.45.0",
    "supabase>=2.25.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ml_system"]
